+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 - support fetch the value from UHUGEINT type column.
 - add `DuckDB::Appender#append_uhugeint`.
-- DuckDB::Result supports ARRAY column type.
+- DuckDB::Result supports ARRAY column type (only when DuckDB::Result.use_chunk_each is true).
+- DuckDB::Result supports LIST column type (only when DuckDB::Result.use_chunk_each is true).
 
 # 1.0.0.0 - 2024-06-08
 - bump duckdb to 1.0.0.